+++ impacted-test-engine/src/test/java/com/teamscale/test_impacted/engine/InternalImpactedTestEngineTest.java
@@ -76,7 +76,7 @@ class InternalImpactedTestEngineTest {
 		when(testwiseCoverageAgentApi.testStarted(anyString())).thenReturn(mock(Call.class));
 		when(testwiseCoverageAgentApi.testFinished(anyString())).thenReturn(mock(Call.class));
 		when(testwiseCoverageAgentApi.testFinished(anyString(), any())).thenReturn(mock(Call.class));
-		when(testwiseCoverageAgentApi.testRunFinished()).thenReturn(mock(Call.class));
+		when(testwiseCoverageAgentApi.testRunFinished(any())).thenReturn(mock(Call.class));
 	}
 
 	@SafeVarargs