+++ tia-client/src/main/java/com/teamscale/tia.client/ITestwiseCoverageAgentApi.java
@@ -66,9 +66,12 @@ public interface ITestwiseCoverageAgentApi {
 			@Body List<ClusteredTestDetails> availableTests
 	);
 
-	/** Test run finished. Generate test-wise coverage report and upload to Teamscale. */
+	/**
+	 * Test run finished. Generate test-wise coverage report and upload to Teamscale.
+	 * @param partial Whether the test recording only contains a subset of the available tests.
+	 */
 	@POST("testrun/end")
-	Call<ResponseBody> testRunFinished();
+	Call<ResponseBody> testRunFinished(@Query("partial") Boolean partial);
 
 	/**
 	 * Generates a {@link Retrofit} instance for this service, which uses basic auth to authenticate against the server